Refactor ConsoleView into a View trait
The ConsoleView was refactored into a View trait to make the system more flexible and extendable. Now, different views (e.g., GUI-based) can be implemented and used instead of just a console view. The View trait consists of a render method which was added to the ConsoleView. The render method was also added to the Controller to enable diverse view implementations.