Simplify GUI code and enable grid cell updates
This commit streamlines the code in `src/main/scala/ui/GUIView.scala` by removing many unnecessary imports and code snippets, especially those connected to the initial `text` Label, which has been removed. The commit also introduces a more efficient system for updating the displayed grid cells. Now, instead of creating new Buttons on each frame update, the buttons are stored in an array which is created at startup (`cells`). This array can then be used to directly update the text of the buttons as the grid updates, which should lead to smoother animations and better performance.