Adjust circle radius in GUIView
Modified the calculation of circle's radius in GUIView to be half of 80% of the stack's width instead of being half the stack's width reduced by one. This change is aimed to preventing overlap with other user interface elements and helps fit the circle more appropriately within its designated container space.