Refactor game classes and add new AI strategy
Changed various classes to generalize from GoGame to a more generic Game class, allowing for potential future extensibility and abstraction. In the AI trait and its implementations (FirstMoveAI, RandomAI), the game parameter type has been updated to utilize this Game class. Renamed GoGame to Game and GoApp to App for the same reason.

Added new AI strategy named BestKillingMoveAI which encapsulates logic of finding and executing the move that can kill the most opponent's groups.

Updated the Controller to take an AI instance as a parameter to make it easier to switch between different AI strategies in the future. In the App, created and passed in an instance of BestKillingMoveAI in the main function.