Refactor animations for stone transition
This commit changes animation handling for the game stones. Instead of just having a TranslateTransition, the code now makes use of ParallelTransition that includes FadeTransition along with TranslateTransition. This improves the visual appearance of the stone placement and removal in the UI of the game by making it more smooth and seamless. This change makes the UX more enriched by adding a more visually appealing transition effect.