Add scale transition to GUI animations
In this commit, a scale transition has been added to the user interface animations. Now, alongside the existing fade and translate transitions, UI elements will also scale up or down. This change was implemented to make animations feel more dynamic and immersive, improving the overall user interaction with the interface.