Refactor GUIView for readability and better button initializing
In this commit, layout and organization of the code in GUIView is improved for better readability. Initially, cells are moved to the start function, aligning them with the grid they are added to. Button creation is also modified to allow StackPane buttons for cleaner design and easier user interaction. Other slight efficiency improvements are made such as replacing zipWithIndex with regular for loops for cell addition to the grid.