"Improve AI calculations for opponent's attacks
The code in LessDumbAI has been modified to take into account the potential kills the opponent could make in the next move. Previously, the AI was only considering the stones it could kill in the next move. Now, it's also considering the stones the opponent might kill if it was their turn. This results in smarter decisions of the AI as it's not only thinking about offense, but also defense."