Improve game performance, implement clean exit
Applied optimizations to the application's main loop to reduce CPU usage and removed an unnecessary artificial sleep delay. A severe bottleneck was the delay that was set to 2000ms and it has been removed. Also, the turn throttle has been decreased from 1000 to 10 and the sleep time between turns has been decreased from 500ms to 200ms to speed up the game's execution.

In addition, to prevent processes from lingering after the game has ended, a call to Platform.exit() has been added after a short 5000ms delay. This ensures that all JavaFX threads are terminated properly when the game ends, providing a clean exit.