Update GUIView to dynamically accommodate different grid sizes
Grid sizes in the GUIView were hardcoded to 9, therefore it was limiting the game to a 9x9 grid. We have now included a gridSize parameter in the GUIView class and replaced all hardcoded grid sizes with gridSize.getSize to support custom game board sizes. This change allows us to dynamically update the visuals for different game board sizes. This was also replicated in App.scala, removing the hardcoded initialization of the Game instance and replacing it with a gridSize parameter.