Add functionality for human player and improve input handling
A Human.scala and Player.scala file have been added to allow the addition of human players into the game, and to outline a structure for players in general. The play method within the Human class handles user input for moves, and checks if these moves are valid. Additionally, the AI.scala file has been modified to extend the Player trait, now requiring an implementation of the play method. Move.scala and Game.scala have also been updated to improve the handling of move input and validation. This enhances the user experience by providing feedback on invalid moves and provides tools for further expansion of player types.